+++ src/movie/models/movie_model.py
@@ -1,10 +1,11 @@
 class Movie:
-    def __init__(self, id, title, genre, duration, release_date):
+    def __init__(self, id, title, genre, duration, release_date, image_url=None):
         self.id = id
         self.title = title
         self.genre = genre
         self.duration = duration
         self.release_date = release_date
+        self.image_url = image_url
 
     def to_dict(self):
         return {
@@ -12,7 +13,8 @@ class Movie:
             "title": self.title,
             "genre": self.genre,
             "duration": self.duration,
-            "release_date": self.release_date
+            "release_date": self.release_date,
+            "image_url": self.image_url
         }
 
 class Showtime: